+++ packages/anyhoo_core/lib/anyhoo_core.dart
@@ -4,7 +4,7 @@
 // ignore: unnecessary_library_name
 library anyhoo_core;
 
-export 'models/auth_user.dart';
+export 'models/anyhoo_user.dart';
 export 'models/arguments.dart';
 export 'arguments_parser.dart';
 export 'widgets/error_page.dart';