+++ packages/anyhoo_image_selector/lib/widgets/anyhoo_image_selector_widget.dart
@@ -1,10 +1,10 @@
 import 'dart:typed_data';
 
-import 'package:image_selector/cubit/image_selector_state.dart';
-import 'package:image_selector/cubit/show_stock_photos_cubit.dart';
-import 'package:image_selector/layout_type.dart';
-import 'package:image_selector/widgets/add_stock_photo_page.dart';
-import 'package:image_selector/selected_image.dart';
+import 'package:anyhoo_image_selector/cubit/image_selector_state.dart';
+import 'package:anyhoo_image_selector/cubit/show_stock_photos_cubit.dart';
+import 'package:anyhoo_image_selector/layout_type.dart';
+import 'package:anyhoo_image_selector/widgets/add_stock_photo_page.dart';
+import 'package:anyhoo_image_selector/selected_image.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:image_picker/image_picker.dart';
@@ -17,7 +17,7 @@ const emptyIconTotalSize = 100.0;
 const horizontalSplitSpacing = 16.0;
 const horizontalSplitPadding = 16.0;
 
-class ImageSelectorWidget extends StatelessWidget {
+class AnyhooImageSelectorWidget extends StatelessWidget {
   // Callback when a file is selected
   final ValueChanged<SelectedImage>? onImageSelected;
   final List<String> stockAssetPaths;
@@ -25,7 +25,7 @@ class ImageSelectorWidget extends StatelessWidget {
   final bool roundImage;
   final LayoutType layoutType;
 
-  const ImageSelectorWidget({
+  const AnyhooImageSelectorWidget({
     super.key,
     this.onImageSelected,
     this.stockAssetPaths = const [],