+++ packages/anyhoo_router/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.4
+
+* Change class names
+
 ## 0.0.3
 
 * Revert to have references using path